Uploaded free form responses analysis
This analysis was done by Ben Marwick. An e-mail from him:

I had a bit of a look at the free responses, and tried some bigram analysis, sentiment analysis, word clouds and topic models. I followed closely the examples in the excellent book https://www.tidytextmining.com  There are lots of good ideas in that book for other types of analysis that we could do with this data. It's a great read with good code examples. I've attached my code and output so you can take a look and run it yourself. 

I think you can get some useful insights from these analysis, for example the attitude of the instructors is a strong feature of the positive reactions. Negative reactions relate to getting lost or stuck. Accessibility concerns relate to having enought time, and pacing of the lessons. 

Although we can get some insights from these results, it will be best to back up these observations with specific examples from the free responses. I checked the raw data bunch of times while working on this to ensure that the topics I was seeing in the topic model output was really in people's responses.